Moving code from `function/mod.rs` to `function/attr.rs`.
